mv parser into crate tests
