+++ consensus/scp/Cargo.toml
@@ -19,8 +19,6 @@ mc-util-metrics = { path = "../../util/metrics" }
 mc-util-serial = { path = "../../util/serial", features = ["std"] }
 
 bigint = "4.4"
-pest = "2.0"
-pest_derive = "2.0"
 rand = "0.7"
 rand_hc = "0.2"
 serde = { version = "1.0", default-features = false, features = ["alloc", "derive"] }
@@ -33,6 +31,8 @@ mc-util-logger-macros = { path = "../../util/logger-macros" }
 mc-util-test-helper = { path = "../../util/test-helper" }
 
 crossbeam-channel = "0.3"
+pest = "2.0"
+pest_derive = "2.0"
 pretty_assertions = "0.6.1"
 serial_test = "0.1"
 serial_test_derive = "0.1"
