+++ consensus/scp/src/lib.rs
@@ -3,10 +3,8 @@
 #![feature(external_doc)]
 #![doc(include = "../README.md")]
 #![allow(non_snake_case)]
-// #![deny(missing_docs)]
+#![deny(missing_docs)]
 
-#[macro_use]
-extern crate pest_derive;
 
 pub mod core_types;
 pub mod msg;
